pyproject.toml
[project]
name = "ccbert"
version = "0.1.0"
description = "Official repository for the paper: Designing Cell–Cell Relation Extraction Models: A Systematic Evaluation of Entity Representation and Pre-training Strategies"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "altair>=6.0.0",
    "autopep8>=2.3.2",
    "captum>=0.8.0",
    "cupy-cuda12x>=13.6.0",
    "dask>=2025.12.0",
    "datasets>=4.4.2",
    "deepchem>=2.8.0",
    "dvc>=3.64.2",
    "einops>=0.8.1",
    "evaluate>=0.4.6",
    "fastparquet>=2025.12.0",
    "flake8>=7.3.0",
    "graphviz>=0.21",
    "huggingface-hub>=0.36.0",
    "leidenalg>=0.11.0",
    "lightgbm>=4.6.0",
    "lime>=0.2.0.1",
    "matplotlib>=3.10.8",
    "mlflow>=3.7.0",
    "networkx>=3.6.1",
    "nltk>=3.9.2",
    "numpy>=1.26.4",
    "nvidia-ml-py3>=7.352.0",
    "openai>=2.14.0",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "peft>=0.18.0",
    "pipdeptree>=2.30.0",
    "plotly>=6.5.0",
    "polars>=1.36.1",
    "pubchempy>=1.0.5",
    "pyarrow>=22.0.0",
    "pysqlite3>=0.5.4",
    "python-dateutil>=2.9.0.post0",
    "python-igraph>=1.0.0",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
    "rdkit>=2025.9.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.1",
    "slackweb>=1.0.5",
    "spacy>=3.8.11",
    "tensorboard>=2.20.0",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "voyager>=2.1.0",
    "wandb>=0.23.1",
    "xgboost>=3.1.2",
]
